Ajustes de layout e fluxo do projeto

+++ app/src/main/java/com/example/custofrete/domain/model/CustoViagem.kt
@@ -1,8 +1,8 @@
 package com.example.custofrete.domain.model
 
 data class CustoViagem(
-    val valorAlimentacao:Float,
-    val valorHotel:Float,
-    val valorGasolina:Float,
-    val gastosExtras:Float
+    val valorAlimentacao:Double,
+    val valorHotel:Double,
+    val valorGasolina:Double,
+    val gastosExtras:Double
 )
